Dialer: Modify the condition of showing VT button for Presence
When Presence is enabled, VT button in call log should show/hide
depending on the capability fetched by Presence client, but the
previously logic is to show VT button when it had video call.

So it is modified that when Presence is enabled, VT button depends
on capability of remote party, and when it's disabled, VT button
shows when it had video call.

CRs-Fixed: 1006517
Change-Id: Ib8db0ee5cf33e24aae5960c683bbdb169235e653