IMS-VT: Handle configuration changes on low battery Video call
If UE orientation changes with low battery dialog showing, then
low battery dialog is dismissed and <mAlert> is set to NULL.
This dialog is later shown to user when InCallActivity is
is recreated. At this time, onDimiss of AlertDialog is
getting called that is bringing down the dialog. Fix this
by not setting <mAlert> to NULL in onDismiss handler of
AlertDialog

Also remove code done with <onConfigurationChanging> API that
handles configuration changes on low battery dialog since this
is implicitly handled with <onDestroyInCallActivity> API.

Change-Id: I74c272b22dcb88e2e741690e2cf46af3ebf36f94
CRs-Fixed: 1024387