IMS-VT: Fix race condition issue of no preview during VT call.
Sometimes camera is enabled and immediately disabled due to
onUiShowing callback. Due to the camera capabilities query
during initial camera enable request, camera dimensions are
received while camera is in OFF state. This causes
mPreviewSurfaceState to be set to something other than NONE.
Next time when UI is resumed, and onUiShowing callback is
invoked, camera is not enabled since mPreviewSurfaceState is
not NONE. This causes camera to remain disabled during VT call.
To fix this, register to onUiShowing before onStateChange so
that back to back camera enable and disable is avoided.
Also if camera capabilities are received in camera OFF state,
don't update mPreviewSurfaceState variable.

Change-Id: I1ce18dbc5c1d5f6f5abaa8db7eb14012a59499a1
CRs-Fixed: 2036068